Move the logic to QueryBuilder
